+++ R/roboto-condensed.r
@@ -3,6 +3,11 @@
 #' You should [import_roboto_condensed]() first and also install the fonts on your
 #' system before trying to use this theme.
 #'
+#' There is an option `hrbrthemes.loadfonts` which -- if set to `TRUE` -- will
+#' call `extrafont::loadfonts()` to register non-core fonts with R PDF & PostScript
+#' devices. If you are running under Windows, the package calls the same function
+#' to register non-core fonts with the Windows graphics device.
+#'
 #' @md
 #' @section Why Roboto Condensed?:
 #' It's free, has tolerable kerning pairs and multiple weights. It's also different
@@ -17,6 +22,7 @@
 #' @param caption_family,caption_face,caption_size,caption_margin plot caption family, face, size and margin
 #' @param axis_title_family,axis_title_face,axis_title_size axis title font family, face and size
 #' @param axis_title_just axis title font justificationk one of `[blmcrt]`
+#' @param axis_text_size font size of axis text
 #' @param plot_margin plot margin (specify with [ggplot2::margin])
 #' @param grid_col grid color
 #' @param grid panel grid (`TRUE`, `FALSE`, or a combination of `X`, `x`, `Y`, `y`)
@@ -54,24 +60,25 @@
 #'   theme(axis.text.y=element_blank())
 #' }
 theme_ipsum_rc <- function(
-  base_family="Roboto Condensed", base_size = 11,
+  base_family="Roboto Condensed", base_size = 11.5,
   plot_title_family=base_family, plot_title_size = 18,
   plot_title_face="bold", plot_title_margin = 10,
   subtitle_family=if (.Platform$OS.type == "windows") "Roboto Condensed" else "Roboto Condensed Light",
-  subtitle_size = 12,
+  subtitle_size = 13,
   subtitle_face = "plain", subtitle_margin = 15,
   strip_text_family = base_family, strip_text_size = 12,
   strip_text_face = "plain",
   caption_family=if (.Platform$OS.type == "windows") "Roboto Condensed" else "Roboto Condensed Light",
   caption_size = 9,
   caption_face = "plain", caption_margin = 10,
+  axis_text_size = base_size,
   axis_title_family = base_family,
   axis_title_size = 9,
   axis_title_face = "plain",
   axis_title_just = "rt",
   plot_margin = margin(30, 30, 30, 30),
-  grid_col = "#2b2b2b99", grid = TRUE,
-  axis_col = "#2b2b2b", axis = FALSE, ticks = FALSE) {
+  grid_col = "#cccccc", grid = TRUE,
+  axis_col = "#cccccc", axis = FALSE, ticks = FALSE) {
 
   ret <- ggplot2::theme_minimal(base_family=base_family, base_size=base_size)
 
@@ -80,9 +87,9 @@ theme_ipsum_rc <- function(
 
   if (inherits(grid, "character") | grid == TRUE) {
 
-    ret <- ret + theme(panel.grid=element_line(color=grid_col, size=0.10))
-    ret <- ret + theme(panel.grid.major=element_line(color=grid_col, size=0.10))
-    ret <- ret + theme(panel.grid.minor=element_line(color=grid_col, size=0.05))
+    ret <- ret + theme(panel.grid=element_line(color=grid_col, size=0.2))
+    ret <- ret + theme(panel.grid.major=element_line(color=grid_col, size=0.2))
+    ret <- ret + theme(panel.grid.minor=element_line(color=grid_col, size=0.15))
 
     if (inherits(grid, "character")) {
       if (regexpr("X", grid)[1] < 0) ret <- ret + theme(panel.grid.major.x=element_blank())
@@ -131,8 +138,8 @@ theme_ipsum_rc <- function(
   xj <- switch(tolower(substr(axis_title_just, 1, 1)), b=0, l=0, m=0.5, c=0.5, r=1, t=1)
   yj <- switch(tolower(substr(axis_title_just, 2, 2)), b=0, l=0, m=0.5, c=0.5, r=1, t=1)
 
-  ret <- ret + theme(axis.text.x=element_text(margin=margin(t=0)))
-  ret <- ret + theme(axis.text.y=element_text(margin=margin(r=0)))
+  ret <- ret + theme(axis.text.x=element_text(size=axis_text_size, margin=margin(t=0)))
+  ret <- ret + theme(axis.text.y=element_text(size=axis_text_size, margin=margin(r=0)))
   ret <- ret + theme(axis.title=element_text(size=axis_title_size, family=axis_title_family))
   ret <- ret + theme(axis.title.x=element_text(hjust=xj, size=axis_title_size,
                                                family=axis_title_family, face=axis_title_face))
@@ -160,6 +167,11 @@ theme_ipsum_rc <- function(
 #' Import Roboto Condensed font for use in charts
 #'
 #' Roboto Condensed is a trademark of Google.
+#'
+#' There is an option `hrbrthemes.loadfonts` which -- if set to `TRUE` -- will
+#' call `extrafont::loadfonts()` to register non-core fonts with R PDF & PostScript
+#' devices. If you are running under Windows, the package calls the same function
+#' to register non-core fonts with the Windows graphics device.
 #'
 #' @note This will take care of ensuring PDF/PostScript usage. The location of the
 #'   font directory is displayed after the base import is complete. It is highly