+++ NEWS.md
@@ -7,7 +7,8 @@
 * Added `dict` and `ignore` parameters (with defaults) to `gg_check()` (@karldw) [PR #3].
 * Fixed typos (@benmarwick) [PR #1].
 * Spelling in vignette fixed (@zx8754) [Issue #6].
-* New lighter grid color default (`#cccccc`) [Issue #7].
+* New lighter grid color default `#cccccc` [Issue #7].
+* Added font auto-loading (@alexwhan) [Issue #12].
 
 
 # hrbrthemes 0.1.0